move {Record,RecordView} to the crate root
