+++ wt_mdb/src/lib.rs
@@ -13,6 +13,7 @@ mod session;
 
 use wt_sys::wiredtiger_strerror;
 
+use std::borrow::Cow;
 use std::ffi::{CStr, CString};
 use std::io::ErrorKind;
 use std::num::NonZero;
@@ -119,8 +120,55 @@ impl From<Error> for std::io::Error {
     }
 }
 
+/// A `RecordView` in a WiredTiger table with an i64 key and a byte array value.
+///
+/// The underlying byte array may or may not be owned, the `Record` type alias may be more
+/// convenient when the data is owned.
+#[derive(Clone, Debug, Hash, Eq, PartialEq)]
+pub struct RecordView<'a> {
+    key: i64,
+    value: Cow<'a, [u8]>,
+}
+
+impl<'a> RecordView<'a> {
+    /// Create a new `RecordView` from a key and an unowned byte array value.
+    pub fn new<V>(key: i64, value: V) -> Self
+    where
+        V: Into<Cow<'a, [u8]>>,
+    {
+        RecordView {
+            key,
+            value: value.into(),
+        }
+    }
+
+    /// Return the key.
+    pub fn key(&self) -> i64 {
+        self.key
+    }
+
+    /// Return the value.
+    pub fn value(&self) -> &[u8] {
+        self.value.as_ref()
+    }
+
+    /// Ensure that this RecordView owns the underlying value.
+    pub fn to_owned(self) -> Record {
+        Record::new(self.key(), self.value.to_vec())
+    }
+
+    /// Returns the inner value within the `RecordView`.
+    pub fn into_inner_value(self) -> Cow<'a, [u8]> {
+        self.value
+    }
+}
+
+/// An alias for `RecordView` with `'static` lifetime, may be more convenient when the value is
+/// actually owned.
+pub type Record = RecordView<'static>;
+
 pub use connection::Connection;
-pub use record_cursor::{Record, RecordCursor, RecordView};
+pub use record_cursor::RecordCursor;
 pub use session::Session;
 pub type Result<T> = std::result::Result<T, Error>;
 
@@ -146,8 +194,7 @@ mod test {
     use crate::{
         connection::Connection,
         options::{ConnectionOptions, ConnectionOptionsBuilder},
-        record_cursor::{Record, RecordView},
-        Error, WiredTigerError,
+        Error, Record, RecordView, WiredTigerError,
     };
 
     fn conn_options() -> Option<ConnectionOptions> {
